drop_duplicates for series
