+++ tests/pandas/sr_drop_duplicates.py
@@ -1,0 +1,28 @@
+# Copyright 2021 NVIDIA Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import pandas as pd
+
+from legate import pandas as lp
+from tests.utils import equals
+
+s = pd.Series(["A", "B", "C", "A", "D"])
+ls = lp.Series(s)
+
+for keep in ["first", "last", False]:
+    out_s = s.drop_duplicates(keep=keep).sort_index()
+    out_ls = ls.drop_duplicates(keep=keep).sort_index()
+
+    assert equals(out_ls, out_s)
